+++ ios/Classes/model/error/MapViewLoadFailed.swift
@@ -1,4 +1,4 @@
 internal struct MapViewLoadFailed : BaseError {
-    var errorCode: Int64 = 0
+    var errorCode: Int = 0
     var message: String? = "Map Engine creation or initialization failed."
 }
